VULCAN-18 Remove the need to use a Pagination instance in order to create a `Page`. Thanks to this, developers don't need to specify the model`s class
For this end, create the `PageItems` class, with hold all the necessary information the `RootEndpoint` needs to create the `Page` itself.

As a result of this change the method `createPage` inside `Pagination` is no longer needed.
